Refactor unify OnEditTextListener  in DateRows
