+++ core/src/main/java/org/hisp/dhis/android/sdk/ui/adapters/rows/dataentry/IncidentDatePickerRow.java
@@ -34,7 +34,6 @@ import static android.text.TextUtils.isEmpty;
 import android.app.DatePickerDialog;
 import android.content.Context;
 import android.support.v4.app.FragmentManager;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -50,13 +49,8 @@ import org.hisp.dhis.android.sdk.ui.fragments.dataentry.RowValueChangedEvent;
 import org.joda.time.DateTime;
 import org.joda.time.LocalDate;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Locale;
 
-
-public class IncidentDatePickerRow extends AbsEnrollmentDatePickerRow {
+public class IncidentDatePickerRow extends AbsDatePickerRow {
     private static final String TAG = "IncidentDatePickerRow";
     private Enrollment mEnrollment;
     private String mLabel;
@@ -105,7 +99,7 @@ public class IncidentDatePickerRow extends AbsEnrollmentDatePickerRow {
 
     @Override
     public int getViewType() {
-        return super.getViewType();
+        return DataEntryRowTypes.ENROLLMENT_DATE.ordinal();
     }
 
     private class DatePickerRowHolder {
@@ -124,7 +118,8 @@ public class IncidentDatePickerRow extends AbsEnrollmentDatePickerRow {
 //            this.detailedInfoButton = detailedInfoButton;
 
             dateSetListener = new DateSetListener(pickerInvoker);
-            invokerListener = new OnEditTextClickListener(context, dateSetListener, pickerInvoker);
+            invokerListener = new OnEditTextClickListener(context, dateSetListener,
+                    false, pickerInvoker);
             clearButtonListener = new ClearButtonListener(pickerInvoker);
 
             clearButton.setOnClickListener(clearButtonListener);
@@ -153,38 +148,7 @@ public class IncidentDatePickerRow extends AbsEnrollmentDatePickerRow {
 
     }
 
-    private static class OnEditTextClickListener implements View.OnClickListener {
-        private final Context context;
-        private final DateSetListener listener;
-        private TextView dateText;
 
-        public OnEditTextClickListener(Context context,
-                DateSetListener listener, TextView dateText) {
-            this.context = context;
-            this.listener = listener;
-            this.dateText = dateText;
-        }
-
-        @Override
-        public void onClick(View view) {
-            Calendar calendar = Calendar.getInstance();
-            SimpleDateFormat simpleDateFormat = new SimpleDateFormat(DATE_FORMAT, Locale.US);
-            if (!dateText.getText().toString().isEmpty()) {
-                try {
-                    calendar.setTime(
-                            simpleDateFormat.parse(dateText.getText().toString()));
-                } catch (ParseException e) {
-                    Log.e(TAG, "Invalid date format, can't parse to put in the picker");
-                    e.printStackTrace();
-                }
-            }
-            DatePickerDialog picker = new DatePickerDialog(context, listener,
-                    calendar.get(Calendar.YEAR), calendar.get(Calendar.MONTH),
-                    calendar.get(Calendar.DAY_OF_MONTH));
-            picker.getDatePicker().setMaxDate(DateTime.now().getMillis());
-            picker.show();
-        }
-    }
     private static class ClearButtonListener implements View.OnClickListener {
         private final TextView textView;
         private Enrollment enrollment;
